+++ addon/objects/uml-primitives/fd-uml-sequence-object.js
@@ -3,12 +3,11 @@
 */
 
 import { computed } from '@ember/object';
-import { isArray } from '@ember/array';
-import { isEmpty } from '@ember/utils';
+import $ from 'jquery';
 import joint from 'npm:jointjs';
 
 import FdUmlElement from './fd-uml-element';
-import { SequenceActor } from './fd-uml-sequence-actor';
+import { BaseObject } from './fd-uml-baseobject';
 
 /**
   An object that describes a Sequence Object element on the UML diagram.
@@ -32,9 +31,10 @@ export default FdUmlElement.extend({
     @method JointJS
   */
   JointJS() {
-    let properties = this.getProperties('id', 'name', 'size', 'position');
+    let properties = this.getProperties('id', 'size', 'position');
+    properties.objectModel = this;
+    
     return new SequenceDiagramObject(properties);
-
   },
 });
 
@@ -47,58 +47,112 @@ export default FdUmlElement.extend({
   @namespace flexberry.uml
   @constructor
 */
-export let SequenceDiagramObject = SequenceActor.define('flexberry.uml.sequencediagramObject', {
-  attrs: {
-    size: { 'width': 40, 'height': 40 },
-    rect: { width: 40, height: 40, fill: '#FFFFFF', stroke: 'black' },
-    text: {
-      'ref': 'rect',
-      'ref-y': 0.5,
-      'ref-x': 0.5,
-      'text-anchor': 'middle',
-      'y-alignment': 'middle',
-    }
-  },
-  heightPadding: 20,
+export let SequenceDiagramObject = BaseObject.define('flexberry.uml.sequencediagramObject', {
+  // Minimum width.
+  minWidth: 40,
+
+  // Minimum height.
+  minHeight: 40,
+
+  // Inputs padding by Y.
+  heightPadding: 10,
+
+  // Minimum line lenght.
+  minLineLength: 10,
 }, {
-  initialize: function () {
-    this.on('change', function () {
-      this.updateRectangles();
-      this.trigger('uml-update');
-    }, this);
-    joint.shapes.basic.Generic.prototype.initialize.apply(this, arguments);
+  markup: [
+    '<g class="rotatable">',
+    '<g class="scalable">',
+    '<g class="flexberry-uml-header-rect">',
+    '<rect width="125" height="40" />',
+    '<path d="M 63 40 63 110" />',
+    '<rect x="0" y="0" width="125" height="110" fill="transparent" stroke="transparent"/>',
+    '</g>',
+    '</g>',
+    '</g>'
+  ].join(''),
+
+  getRectangles() {
+    return [
+      { type: 'header', element: this }
+    ];
   },
+});
 
-  getObjName: function () {
-    let ret = this.get('name');
-    if (isEmpty(ret)) {
-      return '';
-    } else {
-      return ret;
-    }
-  },
+export let SequencediagramObjectView = joint.shapes.flexberry.uml.BaseObjectView.extend({
+  template: [
+    '<div class="uml-class-inputs">',
+    '<textarea class="class-name-input header-input" value="" rows="1" wrap="off"></textarea>',
+    '<div class="input-buffer"></div>',
+    '</div>'
+  ].join(''),
 
-  updateRectangles: function () {
-    let attrs = this.get('attrs');
-    let objName = this.getObjName();
-    let lines = isArray(objName) ? objName : [objName];
+  updateRectangles: function (resizedWidth, resizedHeight) {
+    const minWidth = this.model.attributes.minWidth;
+    const minHeight = this.model.attributes.minHeight;
+    const widthPadding = this.model.attributes.widthPadding;
+    const heightPadding = this.model.attributes.heightPadding;
+    const minLineLength = this.model.attributes.minLineLength;
+    const oldSize = this.model.size();
 
-    let maxStringChars = 8;
-    lines.forEach(function (line) {
-      if (line.length > maxStringChars) {
-        maxStringChars = line.length;
+    let newHeight = Math.max( resizedHeight || oldSize.height, minHeight)
+    let newWidth = Math.max( resizedWidth || oldSize.width, minWidth)
+
+    let inputs =  this.$box.find('.class-name-input');
+    let $buffer = this.$box.find('.input-buffer');
+
+    inputs.each(function() {
+      let $input = $(this);
+      $buffer.css('font-weight', $input.css('font-weight'));
+      $buffer.text($input.val());
+      $input.width($buffer.width() + widthPadding);
+      
+      if ($input.width() > newWidth) {
+        newWidth = $input.width();
       }
     });
 
-    let hightStep = attrs.text['font-size'];
-    let rectHeight = lines.length * hightStep + this.get('heightPadding');
+    const rect = this.$el.find('.flexberry-uml-header-rect rect')[0];
+    const path = this.$el.find('.flexberry-uml-header-rect path')[0];
 
-    let widthStep = attrs.text['font-size'] / 1.5;
-    let rectWidth = maxStringChars * widthStep + 10;
+    if (!rect || !path) {
+      this.model.resize(newWidth, newHeight);
+      if (this.model.get('highlighted')) {
+        this.unhighlight();
+        this.highlight();
+      }
+      return;
+    }
 
-    attrs.text.text = lines.join('\n');
-    attrs.rect.height = rectHeight;
-    attrs.rect.width = rectWidth;
-    this.resize(rectWidth, rectHeight);
-  }
+    const pathBBox = path.getBBox();
+    const rectBBox = rect.getBBox();
+
+    const rectHeight = inputs[0].offsetHeight + 2 * heightPadding;
+    newHeight = Math.max(newHeight, rectHeight + minLineLength);
+    const transformX = newWidth / rectBBox.width;
+    const transformY = newHeight / (rectBBox.height + pathBBox.height);
+    rect.setAttribute('height', rectHeight / transformY);
+
+    const scalable = this.$el.find('.scalable')[0];
+    const transform = scalable.transform.baseVal.consolidate().matrix;
+    transform.a = transformX;
+    transform.d = transformY;
+
+    inputs.css({
+      top: (rectHeight - inputs[0].offsetHeight) / 2,
+      left: (newWidth - inputs[0].offsetWidth) / 2,
+      position: 'absolute'
+    });
+
+    const d = path.getAttribute('d');
+    path.setAttribute('d', d.replace(/(M)\s(-?\d+(?:\.\d+)?)\s(-?\d+(?:\.\d+)?)/, (_, command, x) => `${command} ${x} ${rectHeight / transformY}`));
+ 
+    this.model.resize(newWidth, newHeight);
+    if (this.model.get('highlighted')) {
+      this.unhighlight();
+      this.highlight();
+    }
+  },
 });
+
+joint.shapes.flexberry.uml.sequencediagramObjectView = SequencediagramObjectView;