+++ addon/objects/uml-primitives/fd-uml-transition.js
@@ -52,10 +52,6 @@ export default FdUmlBaseLink.extend({
   @constructor
 */
 export let Transition = Connection.define('flexberry.uml.Transition', {
-  attrs: {
-    rect: { 'visibility': 'visible' },
-  }
-}, {
   initialize: function () {
     Dependency.prototype.initialize.apply(this, arguments);
   }